requirements.txt
fal-client
flask
flask-cors
pillow
requests
